testing updates

+++ FilePartitionerTest/WritePartitionsTest.cs
@@ -56,10 +56,13 @@ namespace FilePartitionerTest
             var directory = string.Join("\\", processParts.Take(processParts.Length - 1));
 
             //clean out the files
-            var files = Directory.GetFiles(directory, "*.csv");
-            foreach (var file in files)
+            if (directory != "")
             {
-                File.Delete(file);
+                var files = Directory.GetFiles(directory, "*.csv");
+                foreach (var file in files)
+                {
+                    File.Delete(file);
+                }
             }
 
             var filebase = "file";
